remove editRelations from configuration routes
